[Fix #5782] Fix mininum Ruby target version for `Lint/UnifiedInteger`
Fixes #5782.

This PR prevents autocorrected code from being broken in the following
cases when using Ruby 2.3 or lower.

```ruby
test = 1
if test.class == Fixnum
  # do something
end
```

And this PR adds tests of Ruby 2.3 or lower and changes documents
slightly.

It has been made to support Ruby 2.4 or higher as an intention to
emulate Ruby's warning.

[IMO] `Fixnum` and `Bignum` inherit `Integer`. It should be programmed
to replaced by `Integer` inherently.